Patches to modify texts inside the config wizzard
Some texts are not available via the site-translation feature. But we still want to change them.
Before v2017.1 we used forked gluon-packages to do so. But those did not get any love.
Now we are trying patches that are likey to break if the modules change. This way we can react to problems arising from our patches.